# Run search results through AnnotationJSONPresenter