To allow AnnotationJSONPresenter to be used to manipulate clients' view of annotation data, it needs to be used consistently across all the endpoints that render annotations as JSON. This commit ensures that
annotations returned from search views are presented.

This does have a performance penalty. My testing suggests that it makes the search endpoint approximately 2× slower:

    before
    ------

    $ wrk -c1 -d30 -t1 https://localhost:5000/api/search
    Running 30s test @ https://localhost:5000/api/search
      1 threads and 1 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    25.23ms   10.85ms 131.04ms   96.43%
        Req/Sec    35.10      6.81    40.00     93.31%
      1015 requests in 30.03s, 39.50MB read
    Requests/sec:     33.79
    Transfer/sec:      1.32MB

    after
    -----

    $ wrk -c1 -d30 -t1 https://localhost:5000/api/search
    Running 30s test @ https://localhost:5000/api/search
      1 threads and 1 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    53.09ms    9.46ms 153.72ms   97.07%
        Req/Sec    17.30      4.18    20.00     79.79%
      509 requests in 30.05s, 18.90MB read
    Requests/sec:     16.94
    Transfer/sec:    643.90KB